# openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Cliftin API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/exercises": {
      "get": {
        "operationId": "listExercises",
        "summary": "List exercise metadata",
        "tags": [
          "exercises"
        ],
        "description": "Returns non-deleted exercise definitions. Performance statistics are exposed separately. Use an opaque cursor for pagination.",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false,
            "description": "Opaque cursor returned by a previous page."
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "equipmentId",
            "required": false,
            "description": "Liftin equipment identifier."
          },
          {
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "in": "query",
            "name": "muscle",
            "required": false
          },
          {
            "schema": {
              "minLength": 1,
              "type": "string"
            },
            "example": "bench press",
            "in": "query",
            "name": "q",
            "required": false,
            "description": "Case- and separator-insensitive search over stored, derived, and alternative English names."
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "name"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "-name"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false,
            "description": "Single sort field; - means descending."
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "alternativeEnglishNames": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "defaultProgressMetric": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "equipment": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "equipmentId": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "id": {
                            "minimum": 1,
                            "type": "integer",
                            "example": 1000
                          },
                          "isDeleted": {
                            "type": "boolean"
                          },
                          "isUserCreated": {
                            "type": "boolean"
                          },
                          "name": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "perceptionScale": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "primaryMuscles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "secondaryMuscles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "supports1RM": {
                            "type": "boolean"
                          },
                          "timerBased": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "alternativeEnglishNames",
                          "defaultProgressMetric",
                          "equipment",
                          "equipmentId",
                          "id",
                          "isDeleted",
                          "isUserCreated",
                          "name",
                          "perceptionScale",
                          "primaryMuscles",
                          "secondaryMuscles",
                          "supports1RM",
                          "timerBased"
                        ]
                      }
                    },
                    "nextCursor": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "items"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/exercises/{exerciseId}": {
      "get": {
        "operationId": "getExercise",
        "summary": "Get exercise metadata",
        "tags": [
          "exercises"
        ],
        "description": "Direct lookup can return a soft-deleted exercise with isDeleted true. Statistics and performances remain separate.",
        "parameters": [
          {
            "schema": {
              "minimum": 1,
              "type": "integer"
            },
            "example": 1000,
            "in": "path",
            "name": "exerciseId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "alternativeEnglishNames": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "defaultProgressMetric": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "equipment": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "equipmentId": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "minimum": 1,
                      "type": "integer",
                      "example": 1000
                    },
                    "isDeleted": {
                      "type": "boolean"
                    },
                    "isUserCreated": {
                      "type": "boolean"
                    },
                    "name": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "perceptionScale": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "primaryMuscles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "secondaryMuscles": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "supports1RM": {
                      "type": "boolean"
                    },
                    "timerBased": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "alternativeEnglishNames",
                    "defaultProgressMetric",
                    "equipment",
                    "equipmentId",
                    "id",
                    "isDeleted",
                    "isUserCreated",
                    "name",
                    "perceptionScale",
                    "primaryMuscles",
                    "secondaryMuscles",
                    "supports1RM",
                    "timerBased"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "issues": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "properties": {
                          "message": {
                            "type": "string"
                          },
                          "path": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "message",
                          "path"
                        ]
                      }
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 400,
                      "type": "integer"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": []
}